refactor: 애플리케이션의 설정을 확인할 수 있는 AppConfig 인터페이스 추가, DispatcherServletConfigurator를 mvc 하위로 이동
