+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -2,6 +2,7 @@ package com.techcourse;
 
 import com.interface21.web.WebApplicationInitializer;
 import com.interface21.webmvc.servlet.mvc.tobe.dispatcherservlet.DispatcherServlet;
+import com.interface21.webmvc.servlet.mvc.tobe.dispatcherservlet.DispatcherServletConfigurator;
 import com.interface21.webmvc.servlet.mvc.tobe.handleradapter.HandlerAdapterRegistry;
 import com.interface21.webmvc.servlet.mvc.tobe.handlermapping.HandlerMappingRegistry;
 import jakarta.servlet.ServletContext;
@@ -21,7 +22,8 @@ public class DispatcherServletInitializer implements WebApplicationInitializer {
     @Override
     public void onStartup(final ServletContext servletContext) {
         final HandlerAdapterRegistry handlerAdapterRegistry = DispatcherServletConfigurator.handlerAdapterRegistry();
-        final HandlerMappingRegistry handlerMappingRegistry = DispatcherServletConfigurator.handlerMappingRegistry();
+        final HandlerMappingRegistry handlerMappingRegistry = DispatcherServletConfigurator.handlerMappingRegistry(
+                new TechcourseAppConfig());
 
         final var dispatcherServlet = new DispatcherServlet(handlerMappingRegistry, handlerAdapterRegistry);
