+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/dispatcherservlet/AppConfig.java
@@ -1,0 +1,11 @@
+package com.interface21.webmvc.servlet.mvc.tobe.dispatcherservlet;
+
+import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import java.util.Map;
+
+public interface AppConfig {
+
+    Map<String, Controller> getManualControllers();
+
+    String getControllerBasePackage();
+}
