+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/dispatcherservlet/DispatcherServletConfigurator.java
@@ -1,4 +1,4 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.tobe.dispatcherservlet;
 
 import com.interface21.webmvc.servlet.mvc.tobe.handleradapter.ControllerHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.handleradapter.HandlerAdapterRegistry;
@@ -10,13 +10,13 @@ import com.interface21.webmvc.servlet.mvc.tobe.handlermapping.ManualHandlerMappi
 
 public class DispatcherServletConfigurator {
 
-    public static HandlerMappingRegistry handlerMappingRegistry() {
+    public static HandlerMappingRegistry handlerMappingRegistry(final AppConfig appConfig) {
         final HandlerMappingRegistry handlerMappingRegistry = new HandlerMappingRegistry();
 
-        final HandlerMapping manualHandlerMapping = new ManualHandlerMapping(
-                ManualControllerConfigurator.getManualControllers());
+        final HandlerMapping manualHandlerMapping = new ManualHandlerMapping(appConfig.getManualControllers());
         handlerMappingRegistry.register(manualHandlerMapping);
-        final HandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping("com.techcourse.controller");
+        final HandlerMapping annotationHandlerMapping = new AnnotationHandlerMapping(
+                appConfig.getControllerBasePackage());
         handlerMappingRegistry.register(annotationHandlerMapping);
 
         return handlerMappingRegistry;
